Improved: code to handle the app link creation(legacy or new) based on permissions when redirecting from login page
